后端代码/requirements.txt
ansicon==1.89.0
arrow==1.2.2
asgiref==3.5.0
astroid==2.5.1
backports.zoneinfo==0.2.1
blessed==1.19.1
colorama==0.4.4
Django==2.2
django-cors-middleware==1.5.0
django-picklefield==3.0.1
django-q==1.3.9
django-redis-cache==3.0.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
greenlet==1.1.2
isort==5.8.0
jinxed==1.1.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysql-connector-python==8.0.24
mysqlclient==2.1.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pylint==2.7.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
